refactor: split out MCP authentication into its own file
Just a cleanup
