+++ crates/agentgateway/src/llm/conversion/bedrock.rs
@@ -2415,7 +2415,6 @@ impl ConverseResponseAdapter {
 			})
 			.unwrap_or_default();
 
-		tracing::error!("howardjohn: {:#?}", usage);
 		completions::Response {
 			id: format!("bedrock-{}", chrono::Utc::now().timestamp_millis()),
 			object: "chat.completion".to_string(),
