+++ crates/agentgateway/src/mcp/mod.rs
@@ -1,3 +1,4 @@
+mod auth;
 mod handler;
 mod mergestream;
 mod rbac;
